% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{library_dplyr}
\alias{library_dplyr}
\alias{library_tidyr}
\title{Load \{dplyr\} or \{tidyr\} without functions ending with an underscore}
\usage{
library_dplyr(..., exclude)

library_tidyr(..., exclude)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[base:library]{base::library()}}}

\item{exclude}{A list of functions to exclude. Leave this argument missing to
exclude all underscore functions from the package by default.}
}
\value{
The list of attached packages invisibly, or \code{TRUE}/\code{FALSE} to
indicate success if \code{logical.return = TRUE} is indicated.
}
\description{
This is the same as \code{library(dplyr)} or \code{library(tidyr)}, but excluding all
functions that end with an underscore and that may conflict with \{svTidy\}
corresponding ones. Note that these functions are deprecated in recent
versions of \{dplyr\} and \{tidyr\}, and they are even defunct in version
1.2.0 or greater of \{dplyr\}.
}
\examples{
library_dplyr()
library_tidyr()
search()
# However, the functions with underscore are not directly accessible
# (unless you make library(svTidy) of course)
get0('mutate_')
}
\seealso{
\code{\link[base:library]{base::library()}}
}
