% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciviews_utils.R
\name{list_sciviews_functions}
\alias{list_sciviews_functions}
\title{List SciViews Functions}
\usage{
list_sciviews_functions()
}
\description{
Give a comprehensive list of the SciViews functions (ending with
\verb{_} and similar to their dplyr or tidyr equivalents). There are two major
differences from the original dplyr and tidyr functions: (1) the data-dot
mechanism replaces \code{.} as \verb{.data=} in case it is not provided, and (2) they
use a formula-masking mechanism instead of data-masking or tidy-select. In
this case, you can either use standard evaluation, specifying \code{df$var} like
for many base R function, of a formula, like \code{~var}.
}
