% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-svBase.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{f_}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{svBase}{\code{\link[svBase]{f_}}}
}}

