% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svTidy-package.R
\docType{package}
\name{svTidy-package}
\alias{svTidy}
\alias{svTidy-package}
\title{svTidy: 'SciViews::R' - Tidy Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

SciViews equivalent functions of 'dplyr' and 'tidyr', but faster and using a standard evaluation of arguments or formulas.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/svTidy}
  \item \url{https://www.sciviews.org/svTidy/}
  \item \url{https://sciviews.r-universe.dev/svTidy}
  \item Report bugs at \url{https://github.com/SciViews/svTidy/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
